Create `object` module.
