+++ src/prelude.rs
@@ -1,4 +1,5 @@
 //! The Neon "prelude," a re-exported collection of the most commonly-used Neon APIs.
 
-pub use value::{Handle, JsBuffer, JsArrayBuffer, BinaryData, Class, JsError, ErrorKind, Value, JsValue, JsUndefined, JsNull, JsBoolean, JsString, ToJsString, JsNumber, JsObject, Object, JsArray, JsFunction, Borrow, BorrowMut};
+pub use value::{Handle, JsBuffer, JsArrayBuffer, BinaryData, JsError, ErrorKind, Value, JsValue, JsUndefined, JsNull, JsBoolean, JsString, ToJsString, JsNumber, JsObject, JsArray, JsFunction, Borrow, BorrowMut};
+pub use object::{Object, Class};
 pub use vm::{VmResult, JsResult, JsResultExt, CallKind, Context, ModuleContext, ExecuteContext, ComputeContext, CallContext, FunctionContext, MethodContext};
